fix iterator detection in drawText..
it was not performing a compromise due to encountering the final element
(due to an impossible iterator.end() conditional). now if given a size
greater than we know, but we're at the end of the vector, it'll
compromise and assume it actually meant the last item. so it doesn't
have to be dead nuts, or anything..