save the tilemap to png..works successfully!
it's a bit hard to see anything in the png, but it's binary, and i only
use the red channel for now, which has 256 bits for me to use, which
that will just be for the raw tile type e.g. stone, dirt, etc. etc. all
defined by the main enum. the remaining channels are Blue, Green, and
Alpha, so i still have 768 bits to store *something* in. maybe i'll
store something like enemies or items in the remaining channels... we'll
see. nice to have room for expansion.

The current going size is about ~11 MiB on disk, which is awesome since
it's actually an 8400x2400 image, which is probably the max we'll ever
need, so that's pretty good for a save game of this nature, imho.

granted, this is only using 1 channel presently, and the values range
from 0 to 1 integers for now. so compression probably plays a huge role
in that of course. still. i'm pretty happy with this idea