increase renderable objects..we actually need a lot
but as you can see, it's way to slow to do any of this on the cpu. to
the gpu we go! time to experiment with GLSL and get it to load a tilemap
we give it, where each tile is a pixel, then render the tile using a
spritesheet to a texture, and blitting that...