fix asset loading
honestly..i'm not sure where to go with librocket.

it seems very useful and handy, but as soon as i want to render a
texture that is obtained at runtime, say for a hotbar inventory hud
thing (think, 6 slots or so for quick access)...i don't see how to do
that easily. i thought about making a custom Element, and maybe that
would work. but i wouldn't know how to pass the texture.
it all feels very nasty to me..i think i'm better off writing my own.

it shouldn't be too hard for what i'm trying to do, anyways