shaderless tilemap rendering works..
there's good news and bad news. the good news is that it is not my
fragment shader that is at fault. the bad news, is that i know
the issue and am not entirely sure how to solve it. the issue really
was as simple as "the resolution y isn't divisble by TILE_SIZE"
facepalm