create a vector of entities, add the player to that
in theory, this will allow us to run collision easily between entities
(there's no spatial hashing though, so here's hoping it will be fast
without it..at least for now), and collision against the tilemap.

also, make renderable::render virtual. it's used for debug drawing,
but having a vector<Entity*> breaks that with non virtuals, since
it'd require a downcast